backend/requirements.txt
fastapi
uvicorn
pydantic
librosa
numpy
soundfile
python-multipart
jiwer
python-Levenshtein